+++ src/transport/sendmail/error.rs
