+++ src/transport/smtp/r2d2.rs
@@ -1,5 +1,5 @@
-use crate::smtp::error::Error;
-use crate::smtp::{ConnectionReuseParameters, SmtpClient, SmtpTransport};
+use crate::transport::smtp::error::Error;
+use crate::transport::smtp::{ConnectionReuseParameters, SmtpClient, SmtpTransport};
 use r2d2::ManageConnection;
 
 pub struct SmtpConnectionManager {
